- bugfix for calculateRSSIMapping noise mapping (David Eckhoff) - refactored the AnalogueModel initialization, the analogue model parameters will be now read from the parameter map directly in the virtual initFromMap method, added a template method for creation of analogue models (createAnalogModel<_CLASS_NAME_>(params) - refactored the Decider initialization, the decider parameters will be now read from the parameter map directly in the virtual initFromMap method (as in AnalogueModels), added a template method for creation of decider (createDecider<_CLASS_NAME_>(params), this makes the entire initialization step more generic and avoids long parameter lists for the decider (analogue model) constructors - refactored the BaseDecider class, added special handling for collision detections, base decider can handle multiple channels, channel IDLE detection improved - added some new DeciderToPhyInterface functions for more flexibility - UWBIR stuff was adopted to the MiXiM PhyLayer, Decider definitions UWBIRPhy is now derived from PhyLayerBattery and the DeciderUWBIRED is derived from BaseDecider (formaly it was derived from Decider) - adjusted the ieee802154Narrow example - removed the WorldUtility reference from ChannelAccess - adjusted templates for new AnalogueModel/Decider modifications

+++ src/base/phyLayer/MacToPhyInterface.h
@@ -79,11 +79,11 @@ public:
 	 * @brief Returns the length of the phy header in bits.
 	 *
 	 * Since the MAC layer has to create the signal for
-     * a transmission it has to know the total length of
-     * the packet and therefore needs the length of the
-     * phy header.
+	 * a transmission it has to know the total length of
+	 * the packet and therefore needs the length of the
+	 * phy header.
 	 */
-	virtual int getPhyHeaderLength() const = 0;
+	virtual long getPhyHeaderLength() const = 0;
 
 	/** @brief Sets the channel currently used by the radio. */
 	virtual void setCurrentRadioChannel(int newRadioChannel) = 0;
